+++ pkg/webmiddleware/csrf_test.go
@@ -22,6 +22,7 @@ import (
 	"github.com/gorilla/csrf"
 	"github.com/smartystreets/assertions"
 	"github.com/smartystreets/assertions/should"
+	"go.thethings.network/lorawan-stack/v3/pkg/auth"
 )
 
 func TestCSRF(t *testing.T) {
@@ -29,14 +30,24 @@ func TestCSRF(t *testing.T) {
 	authKey := []byte("1234123412341234123412341234123412341234123412341234123412341234")
 	m := CSRF(authKey)
 
-	t.Run("Protects non-idempotent methods", func(t *testing.T) {
+	t.Run("Protects non-idempotent methods when using a Session Token", func(t *testing.T) {
 		r := httptest.NewRequest(http.MethodPost, "/", nil)
+		r.Header.Set("Authorization", "Bearer "+auth.JoinToken(auth.SessionToken, "XXX", "YYY"))
 		rec := httptest.NewRecorder()
 		m(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-
 		})).ServeHTTP(rec, r)
 		res := rec.Result()
 		a.So(res.StatusCode, should.Equal, http.StatusForbidden)
+	})
+
+	t.Run("Allows non-idempotent methods when using an API Key", func(t *testing.T) {
+		r := httptest.NewRequest(http.MethodPost, "/", nil)
+		r.Header.Set("Authorization", "Bearer "+auth.JoinToken(auth.APIKey, "XXX", "YYY"))
+		rec := httptest.NewRecorder()
+		m(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
+		})).ServeHTTP(rec, r)
+		res := rec.Result()
+		a.So(res.StatusCode, should.Equal, http.StatusOK)
 	})
 
 	t.Run("Allows access with valid CSRF token", func(t *testing.T) {